Update llvm/clang/lld/libcxx to the latest upstream versions
llvm and clang have got new patches that are relevant for mingw
and arm64/win.

compiler-rt has changed things, so we can't ask it to build builtins
for i686 any longer, but need to request it for i386 instead - and
clang will expect it to be named i386 even though we're invoking it
with an i686 target.