github: Run libcxx's tests on the built toolchain
This only works on the current main branch, not on the pinned
15.x release branch.

For the linux-cross build tasks, this requires checking out the
tree before doing the "Fetch versions" task; we want to start
populating the output "install" directory with the version file,
but the checkout action wipes the current directory. Therefore,
first check out, then fetch the version.