prepare-cross-toolchain: Copy all runtime DLLs to the main bin directory
We used to copy libc++ and libunwind, since those were needed
to actually run the toolchain - but others, like the asan runtime,
may be needed for the user to be able to run the executables that
are built. If this bin directory is in the user's $PATH, they should
be picked up automatically.

For distribution, and for architectures other than the default,
the user still needs to use the binaries from the arch specific
directories.