github: Try to use ccache for the builds of LLVM
This should allow speeding up successive rebuilds of the same
build configuration, as long as the base LLVM version that is
being built is the same.

The second stage builds are run with

    CCACHE_COMPILERCHECK="%compiler% -v"

as the mtime of the compiler can change in successive rebuilds
of the first stage, even if the output binaries are identical.
(Although we could fix the timestamps when tarring the output.)

On the level of building LLVM, there is no significant measurable
difference in build speed between using
CCACHE_COMPILERCHECK="%compiler% -v" and CCACHE_COMPILERCHECK="none".