github: Try to use ccache for the builds of LLVM
This should allow speeding up successive rebuilds of the same
build configuration, as long as the base LLVM version that is
being built is the same.

The second stage builds are run with

    CCACHE_COMPILERCHECK="%compiler% -v"

as the mtime of the compiler can change in successive rebuilds
of the first stage, even if the output binaries are identical.
(Although we could fix the timestamps when tarring the output.)

On the level of building LLVM, there is no significant measurable
difference in build speed between using
CCACHE_COMPILERCHECK="%compiler% -v" and CCACHE_COMPILERCHECK="none".

The max cache sizes are set quite tight; the total size of stored
caches of 10 GB on github actions doesn't last long, thus the larger
the individual caches are, the lesser the chance is that a suitable
cache (for the right branch etc) is found available.

If we do get a cache found from a previous run, but if it isn't
a suitable match, we end up filling up the size available to
ccache with a separate set of compiled objects, forever bloating
the caches for that branch, until it gets expired or pruned.

Currently (on the 18.x LLVM branch), the total cache usage after
a full build is as follows:

                 Size shown by ccache -s   Compressed GHA cache size
    linux                        214 MB    181 MB
    linux-cross-aarch64          225 MB    189 MB
    macos                        0.4 GB    332 MB
    msys2 mingw64                0.2 GB    154 MB
    msys2 clang64                0.3 GB    226 MB
    linux-cross-windows i686     0.37 GB   331 MB
    linux-cross-windows x86_64   0.38 GB   338 MB
    linux-cross-windows armv7    0.36 GB   325 MB
    linux-cross-windows aarch64  0.37 GB   331 MB

The size as shown by "ccache -s" is what matters wrt what we set
in the max-size field, while the compressed size is what matters
for how much storage the GHA caches takes (regarding pruning old
entries to get down to the 10 GB limit).

Currently, the set of cached entries sum up to 2407 MB - and this
is when each cache only contains one clean run; if we restore an
old cache and there are cache misses, we'll keep the old object
files as well, up to the max-size limit.

Thus, with a clean run, the 10 GB limit can only accommodate 4
full runs in total - if the cache usage isn't perfect, they'll
grow and we probably can keep even fewer entries.