build-cross-tools: Add github actions specific markers around the python build
Building python outputs a very large amount of output - usually
an order of magnitude more output than the rest of the same build
step (building LLVM and some tools).

By adding markers, the github actions log should be able to fold
away this bit, to avoid it swamping the rest of the output.

When not running in github actions, no markers should be printed.