build-mingw-w64: Bump to the latest version as of Sep 17th
This is the latest version, minus one commit.

The very latest commit, 20331109d1c90c19a510408ab119a76004bc41ea,
causes issues with compiling VLC; that is a VLC issue (redefining
GUIDs that now are provided by the system headers), but holding
off of that change until VLC is updated to cope with it.