github: Run test-ffmpeg on all 4 architectures
Test both i686 and x86_64, and armv7 and aarch64 on windows-11-arm
runners.

Building this on arm requires passing the --host option; on x86
the default "uname -m" works for both i686 and x86_64 (ffmpeg's
configure maps this to the architecture "x86", and the pointer
size is tested from the compiler), but for arm architectures we
need to specify them explicitly.