github: Run test-libcxx on ARM too
This requires setting CMAKE_CXX_COMPILER_TARGET manually; the LLVM
build system infers the wrong architecture when CMake is an
x86_64 binary.

(The wrong inferred architecture doesn't affect building libcxx,
but the wrong architecture gets passed as --target= parameter
when running the tests.)

Only run check-cxx on ARM; there are a few unwinding tests
(in check-cxxabi and check-unwind) that fail on ARM, relating to
forced unwinding.

Additionally, a couple tests need fixing in check-cxx; one
category is fixed in the latest git main already (with backport
request ongoing), and the other has been submitted as a fix PR
recently.

Co-authored-by: Jeremy Drake <github@jdrake.com>