github: Build linux-stage1 and linux in parallel on aarch64 and x86_64
This makes the linux aarch64 package natively built on aarch64,
rather than cross compiled from x86_64.

The linux aarch64 runners currently seem to be faster than the
x86_64 ones, but it doesn't affect the total pipelien runtime as
the matrices wait for all jobs in them to complete before proceeding.