WIP: github: Switch to Ubuntu 24.04
Building LLVM for RISC-V on Ubuntu 22.04 fails, due to needing to
link in -latomic in a few cases where it isn't done currently.

(The build of LLVM does include LLVM_ATOMIC_LIB in a bunch of
places, but not everywhere where it is needed, if building with
dylib enabled.)

This isn't an issue if building with Ubuntu 22.04's own CMake,
which is patched to add extra -latomic on riscv, but the github
action runners have their own, more modern (vanilla) CMake,
which doesn't do that.